refactor(components): unify Button code in one file
feature/button-enums

HEAD
