+++ src/components/Button/Button.spec.js
@@ -21,4 +21,87 @@ describe('Button', () => {
     const wrapper = create(<Button>Regular Button</Button>);
     expect(wrapper).toMatchSnapshot();
   });
+
+  it('should have button styles', () => {
+    const actual = create(<Button>Button</Button>);
+    expect(actual).toMatchSnapshot();
+  });
+
+  it('should have kilo button styles', () => {
+    const actual = create(<Button size={Button.KILO}>Button</Button>);
+    expect(actual).toMatchSnapshot();
+  });
+
+  it('should have mega button styles', () => {
+    const actual = create(<Button size={Button.MEGA}>Button</Button>);
+    expect(actual).toMatchSnapshot();
+  });
+
+  it('should have giga button styles', () => {
+    const actual = create(<Button size={Button.GIGA}>Button</Button>);
+    expect(actual).toMatchSnapshot();
+  });
+
+  it('should have disabled button styles', () => {
+    const actual = create(<Button disabled>Disabled button</Button>);
+    expect(actual).toMatchSnapshot();
+  });
+
+  it('should have secondary button styles', () => {
+    const actual = create(<Button secondary>Secondary button</Button>);
+    expect(actual).toMatchSnapshot();
+  });
+
+  it('should have secondary disabled button styles', () => {
+    const actual = create(
+      <Button secondary disabled>
+        Secondary disabled button
+      </Button>
+    );
+    expect(actual).toMatchSnapshot();
+  });
+
+  it('should have stretch button styles', () => {
+    const actual = create(<Button stretch>Stretched button</Button>);
+    expect(actual).toMatchSnapshot();
+  });
+
+  /**
+   * Accessibility tests.
+   */
+  it('should meet accessibility guidelines', async () => {
+    const wrapper = renderToHtml(<Button>Button</Button>);
+    const actual = await axe(wrapper);
+    expect(actual).toHaveNoViolations();
+  });
+
+  /**
+   * Logic tests.
+   */
+  describe('as anchor element', () => {
+    it('should become a link when passed an href prop', () => {
+      const { getByTestId } = render(
+        <div>
+          <Button href="#" data-testid="link-button">
+            Link Button
+          </Button>
+        </div>
+      );
+      const buttonEl = getByTestId('link-button');
+      expect(buttonEl.tagName).toBe('A');
+      expect(buttonEl).toHaveAttribute('href');
+    });
+
+    it('should accept a target prop', () => {
+      const { getByTestId } = render(
+        <div>
+          <Button href="#" target="_blank" data-testid="link-button">
+            Link Button
+          </Button>
+        </div>
+      );
+      const buttonEl = getByTestId('link-button');
+      expect(buttonEl).toHaveAttribute('target', '_blank');
+    });
+  });
 });
